requirements.txt
asgiref==3.2.7
boto3==1.13.24
botocore==1.16.24
click==7.1.2
dicttoxml==1.7.4
dj-database-url==0.5.0
Django==2.2.13
django-heroku==0.3.1
django-storages==1.9.1
document==1.0
docutils==0.15.2
gunicorn==20.0.4
jmespath==0.10.0
lxml==4.5.1
Pillow==7.1.2
psycopg2==2.8.5
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
redis==3.5.3
rq==1.4.2
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
